backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.19

motor==3.6.0
pymongo==4.9.2

passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

httpx==0.27.2
tenacity==9.0.0

litellm==1.55.3

pypdf==5.1.0
python-docx==1.1.2
reportlab==4.2.5

